Plugin Generator Updates
