+++ .gitignore
@@ -5,6 +5,13 @@
 /coverage/
 /doc/
 /pkg/
-/spec/manageiq
 /spec/reports/
 /tmp/
+
+.rubocop-*
+/bundler.d/*
+!/bundler.d/.keep
+/config/settings.local.yml
+/config/settings/*.local.yml
+/config/environments/*.local.yml
+/spec/manageiq
