+++ .rspec
@@ -1,3 +1,5 @@
+--require manageiq/spec/spec_helper
 --require spec_helper
 --color
 --order random
+--exclude-pattern "spec/manageiq/**/*_spec.rb"
