+++ .rspec_ci
@@ -1,4 +1,6 @@
+--require manageiq/spec/spec_helper
 --require spec_helper
 --color
 --order random
 --profile 25
+--exclude-pattern "spec/manageiq/**/*_spec.rb"
