+++ .rubocop.yml
@@ -1,4 +1,3 @@
 inherit_from:
 - https://raw.githubusercontent.com/ManageIQ/guides/master/.rubocop_base.yml
-# put all local rubocop config into .rubocop_local.yml as it will be loaded by .rubocop_cc.yml as well
 - .rubocop_local.yml
