+++ .travis.yml
@@ -1,21 +1,9 @@
----
-dist: xenial
 language: ruby
+cache: bundler
 rvm:
 - 2.5.7
 - 2.6.5
-sudo: false
-cache: bundler
 addons:
   postgresql: '10'
-bundler_args: "--no-deployment"
-before_script: bundle exec rake spec:setup
 install: bin/setup
-after_script: bundle exec codeclimate-test-reporter
-notifications:
-  webhooks:
-    urls:
-    - https://webhooks.gitter.im/e/46abb9c5f1f009b6a699
-    on_success: change
-    on_failure: always
-    on_start: never
+after_script: bin/ci/after_script
