+++ .yamllint
@@ -1,6 +1,7 @@
 ---
 ignore: |
   /vendor/**
+  /spec/manageiq/**
 
 extends: relaxed
