+++ bin/ci/after_script
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+./cc-test-reporter after-build --exit-code $TRAVIS_TEST_RESULT
