+++ bin/update
@@ -4,11 +4,11 @@ require 'pathname'
 gem_root = Pathname.new(__dir__).join("..")
 
 if gem_root.join("spec/manageiq").symlink?
-  puts "== SKIPPING update of spec/manageiq because its symlinked =="
+  puts "== SKIPPING update of spec/manageiq because it is symlinked =="
 else
   puts "== Updating manageiq sample app =="
   system("git pull", :chdir => gem_root.join("spec/manageiq"))
 end
 
 require gem_root.join("spec/manageiq/lib/manageiq/environment").to_s
-ManageIQ::Environment.manageiq_plugin_update
+ManageIQ::Environment.manageiq_plugin_update(gem_root)
