+++ lib/manageiq/content/engine.rb
@@ -1,9 +1,9 @@
-require 'rails/engine'
-
 module ManageIQ
   module Content
     class Engine < ::Rails::Engine
       isolate_namespace ManageIQ::Content
+
+      config.autoload_paths << root.join('lib').to_s
 
       def self.vmdb_plugin?
         true
