+++ manageiq-content.gemspec
@@ -1,22 +1,24 @@
-$:.push File.expand_path("../lib", __FILE__)
+# coding: utf-8
+lib = File.expand_path('../lib', __FILE__)
+$LOAD_PATH.unshift(lib) unless $LOAD_PATH.include?(lib)
+require 'manageiq/content/version'
 
-# Maintain your gem's version:
-require "manageiq/content/version"
+Gem::Specification.new do |spec|
+  spec.name          = "manageiq-content"
+  spec.version       = ManageIQ::Content::VERSION
+  spec.authors       = ["ManageIQ Authors"]
 
-# Describe your gem and declare its dependencies:
-Gem::Specification.new do |s|
-  s.name        = "manageiq-content"
-  s.version     = ManageIQ::Content::VERSION
-  s.authors     = ["ManageIQ Developers"]
-  s.homepage    = "https://github.com/ManageIQ/manageiq-content"
-  s.summary     = "Default ManageIQ content"
-  s.description = "Default ManageIQ content"
-  s.licenses    = ["Apache-2.0"]
+  spec.summary       = "Content plugin for ManageIQ."
+  spec.description   = "Content plugin for ManageIQ."
+  spec.homepage      = "https://github.com/ManageIQ/manageiq-content"
+  spec.license       = "Apache-2.0"
 
-  s.files = Dir["{content,lib}/**/*", "LICENSE.txt", "Rakefile", "README.md"]
+  spec.files         = `git ls-files -z`.split("\x0").reject { |f| f.match(%r{^(test|spec|features)/}) }
+  spec.bindir        = "exe"
+  spec.executables   = spec.files.grep(%r{^exe/}) { |f| File.basename(f) }
+  spec.require_paths = ["lib"]
 
-  s.add_runtime_dependency "savon", "~>2.11.1" # Because users expect it to be there for custom code
+  spec.add_runtime_dependency "savon", "~>2.11.1" # Because users expect it to be there for custom code
 
-  s.add_development_dependency "codeclimate-test-reporter"
-  s.add_development_dependency "simplecov"
+  spec.add_development_dependency "simplecov"
 end
